trust_tone_evaluator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trust-tone-evaluator"
version = "0.1.0"
description = "Tone/prosody analysis module for trust assessment in human-robot interaction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Trust HRI Research"}
]
keywords = ["trust", "human-robot-interaction", "speech", "prosody", "emotion", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.35.0",

    # Audio processing
    "librosa>=0.10.0",
    "opensmile>=2.5.0",
    "soundfile>=0.12.0",
    "resampy>=0.4.0",

    # Data handling
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",

    # Training
    "pytorch-lightning>=2.1.0",
    "wandb>=0.16.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",

    # Evaluation
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Utilities
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "ruff>=0.1.0",
]

[project.scripts]
trust-tone-train = "trust_tone_evaluator.scripts.train:main"
trust-tone-evaluate = "trust_tone_evaluator.scripts.evaluate:main"
trust-tone-predict = "trust_tone_evaluator.scripts.predict:main"

[project.urls]
Homepage = "https://github.com/your-org/trust-tone-evaluator"
Documentation = "https://github.com/your-org/trust-tone-evaluator#readme"
Repository = "https://github.com/your-org/trust-tone-evaluator"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
trust_tone_evaluator = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=trust_tone_evaluator --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
